ci: turn off fail-fast strategy
The failing unit tests can not be reproduced consistently.
So I think we'd better continue running tests in all other environments
and see if it is a universal issue.

If not, we'd better go on with the new Cypress version and try to
resolve the flaky test issue on the Cypress side.

Otherwise we may have to roll back to the previous version.