refactor!: use `@vue/create-eslint-config` for ESLint config generation
Airbnb & Standard options are not added yet.

The difference with the previous implementation is that we now
use `@vue/eslint-config-typescript` instead of
`@vue/eslint-config-typescript/recommended`, because I'm not very
satisfied with the current implementation of the latter config.
There are some edge cases that may confuse users.

I need to find time to refactor that package in the coming months.