fix: specify the sub-project to type-check
Fixes #267

It's because I moved the `compilerOptions` from `tsconfig.json` to
`tsconfig.app.json` in the 2023-05-05 (v3.6.2) release, but forgot to
change the `type-check` script.

This isn't comprehensive because it doesn't check all the sub-projects,
but at least it's the same scope as previous versions.

Ideally I hope #274 can work reliably. But now I'm not sure. So let's
first fix this bug.