feat: replace npm-run-all with npm-run-all2
https://www.npmjs.com/package/npm-run-all2

It is a maintained fork of npm-run-all, which is not maintained anymore.
Both maintainers of the fork also maintain some other popular packages,
so I consider it a safe replacement.

Another popular alternative is [concurrently](https://www.npmjs.com/package/concurrently).
But [its size](https://packagephobia.com/result?p=concurrently@8.2.1) is too big,
I think it would be an overkill.

Closes #306