fix: snapshot generation
It looks like the addition of nightwatch broke the snapshot generation, as `nightwatch` was added to the deny list (correctly),
but it means that the _only_ combinations that are filtered are the ones with the 3 e2e test libraries at the same time.

We also need to filter out the combination of Cypress with Playwright, Cypress with Nightwatch, and Playwright with Nightwatch.

See https://github.com/vuejs/create-vue-templates/ that currently includes cypress-playwright for example.

This commit updates the script to remove all generated combinations, and to generate only the proper ones.