docs: Vue DevTools -> Vue DevTools 7 to reduce confusion
The commonly known Vue DevTools extension is version 6.x, which is
available on the Chrome Web Store.
https://github.com/vuejs/devtools

The experimental option in create-vue is to add the Vue DevTools 7,
which is still under development and not yet available on the
Chrome Web Store, only available as a Vite plugin and a Electron app.
https://github.com/vuejs/devtools-next