feat: add vscode file nesting config to base template
We now have *a lot* of configuration files at the project's root. It would help to have them grouped in the editor.

Many people aren't aware of this feature, so I'm adding a default configuration to the base template to raise awareness.

This is a very rudimentary configuration; only patterns relevant to this project are included. Users can customize it further themselves.

Or they can use a more advanced share configuration like https://github.com/antfu/vscode-file-nesting-config