feat: `--minimal` option
Closes #112
Closes #300

The minimal template is a cut-down version of the default template;
therefore, the `--minimal` flag is mutually exclusive with all other flags.
I chose not to include it in the prompts because it's a very specific
use case, and it's not a good idea to clutter the CLI with too many
options that are not commonly used.

I didn't design it as an add-on because the logic would be too complex,
and the use case would be even more niche.