feat!: use a giant multiselect instead of individual toggles for features
As we are adding more features, the number of toggles is getting out of hand.
A multiselect could save a few keystrokes for most users.

I also take this opportunity to change the prompts library to `@clack/prompts`.
The color library is changed to `picocolors`. IMO `kleur` is still the smaller
library because of treeshaking. But `@clack/prompts` already comes with
`picocolors` as a dependency, so by not adding `kleur`, we save a few bytes.

This is a proof-of-concept change. I haven't put much consideration into the
code style yet. Nor have I updated the i18n messages. I will do that if this
change is accepted.

See the result in a screen recording:
[![asciicast](https://asciinema.org/a/lPaZSGz9LTd7mcWZeJAZxerhx.svg)](https://asciinema.org/a/lPaZSGz9LTd7mcWZeJAZxerhx)