feat: move all eslint-related stuff to main repo
It's becoming harder and harder to maintain `create-eslint-config`
separately as the tooling space evolves. So I decided to merge it back
to the main repo.

The functionality remains mostly the same, except that we now explicitly
generate an oxlint config file instead of passing options via CLI args.
Previously I was adding `--ignore-path .gitignore` to oxlint commands,
but it turns out `.gitignore` is automatically respected inside a git
repo[^1]. So I skipped that part.

The general idea of this refactoring is that each functionality should
be self-contained inside its own template folder, so that in the future
we can test them independently or even publish them as separate packages
again. Still a lot of work to do in that direction, but this is a start.

[^1]: https://oxc.rs/docs/guide/usage/linter/cli.html#ignore-files