feat: move all eslint-related stuff to main repo (#875)
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: Haoqun Jiang <haoqunjiang@gmail.com>
Co-authored-by: Zhousg <zhoushugang@itcast.cn>
Co-authored-by: Thorsten Lünborg <t.luenborg@googlemail.com>
Co-authored-by: Cédric Exbrayat <cexbrayat@users.noreply.github.com>
Co-authored-by: Anton Lilleby <antonlilleby@Anton-sin-MacBook-Pro.local>
Co-authored-by: Dmytro Tsysar <1106994+idmytro@users.noreply.github.com>
Co-authored-by: btea <2356281422@qq.com>

It's becoming harder and harder to maintain `@vue/create-eslint-config` separately as the tooling space evolves. So I decided to merge it back to the main repo.

The functionality remains mostly the same, except that we now explicitly generate an oxlint config file instead of passing options via CLI args.
Previously, I was adding `--ignore-path .gitignore` to oxlint commands, but it turns out `.gitignore` is automatically respected inside a git repo[^1]. So I skipped that part.

The general idea of this refactoring is that each general feature should be self-contained inside its own template folder (`linting` and `formatting` in this PR, in parallel with `tsconfig` and other stuff), so that in the future we can test them independently or even publish them as separate packages again. Still a lot of work to do in that direction, but this is a start.

[^1]: https://oxc.rs/docs/guide/usage/linter/cli.html#ignore-files